+++ src/som/interpreter/actors/EventualSendNode.java
@@ -31,7 +31,8 @@ import som.vm.constants.Nil;
 import som.vmobjects.SSymbol;
 import tools.SourceCoordinate;
 import tools.SourceCoordinate.FullSourceCoordinate;
-import tools.actors.Tags.EventualMessageSend;
+import tools.concurrency.Tags.EventualMessageSend;
+import tools.concurrency.Tags.ExpressionBreakpoint;
 import tools.debugger.nodes.AbstractBreakpointNode;
 import tools.debugger.nodes.BreakpointNodeGen;
 import tools.debugger.nodes.DisabledBreakpointNode;
@@ -289,7 +290,7 @@ public class EventualSendNode extends ExprWithTagsNode {
 
     @Override
     protected boolean isTaggedWith(final Class<?> tag) {
-      if (tag == EventualMessageSend.class) {
+      if (tag == EventualMessageSend.class || tag == ExpressionBreakpoint.class) {
         return true;
       }
       return super.isTaggedWith(tag);