+++ src/som/primitives/processes/ChannelPrimitives.java
@@ -8,11 +8,14 @@ import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
 import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.source.SourceSection;
 
+import som.VM;
 import som.compiler.AccessModifier;
 import som.compiler.MixinBuilder.MixinDefinitionId;
 import som.interpreter.actors.Actor.UncaughtExceptions;
+import som.interpreter.actors.SuspendExecutionNodeGen;
 import som.interpreter.nodes.nary.BinaryComplexOperation;
 import som.interpreter.nodes.nary.TernaryExpressionNode;
 import som.interpreter.nodes.nary.UnaryExpressionNode;
@@ -33,6 +36,15 @@ import som.vmobjects.SInvokable;
 import som.vmobjects.SObject.SImmutableObject;
 import som.vmobjects.SObjectWithClass;
 import som.vmobjects.SSymbol;
+import tools.SourceCoordinate;
+import tools.SourceCoordinate.FullSourceCoordinate;
+import tools.concurrency.Tags.ChannelRead;
+import tools.concurrency.Tags.ChannelWrite;
+import tools.concurrency.Tags.ExpressionBreakpoint;
+import tools.debugger.nodes.AbstractBreakpointNode;
+import tools.debugger.nodes.BreakpointNodeGen;
+import tools.debugger.nodes.DisabledBreakpointNode;
+import tools.debugger.session.Breakpoints;
 
 
 public abstract class ChannelPrimitives {
@@ -58,7 +70,7 @@ public abstract class ChannelPrimitives {
     }
   }
 
-  private static final class ProcessThread extends ForkJoinWorkerThread {
+  public static final class ProcessThread extends ForkJoinWorkerThread {
     ProcessThread(final ForkJoinPool pool) { super(pool); }
   }
 
@@ -119,42 +131,101 @@ public abstract class ChannelPrimitives {
     }
   }
 
-  @Primitive(primitive = "procRead:")
+  @Primitive(primitive = "procRead:", selector = "read")
   @GenerateNodeFactory
   public abstract static class ReadPrim extends UnaryExpressionNode {
-    public ReadPrim(final boolean eagerlyWrapped, final SourceSection source) { super(eagerlyWrapped, source); }
+    /** Halt execution when triggered by breakpoint on write end. */
+    @Child protected UnaryExpressionNode haltNode;
+
+    /** Breakpoint info for triggering suspension after write. */
+    @Child protected AbstractBreakpointNode afterWrite;
+
+    public ReadPrim(final boolean eagerlyWrapped, final SourceSection source) {
+      super(eagerlyWrapped, source);
+      haltNode = SuspendExecutionNodeGen.create(false, sourceSection, null);
+
+      if (VmSettings.TRUFFLE_DEBUGGER_ENABLED) {
+        Breakpoints bpCatalog = VM.getWebDebugger().getBreakpoints();
+        FullSourceCoordinate coord = SourceCoordinate.create(source);
+        this.afterWrite = insert(BreakpointNodeGen.create(bpCatalog.getOppositeBreakpoint(coord)));
+      } else {
+        this.afterWrite = insert(new DisabledBreakpointNode());
+      }
+    }
 
     @Specialization
-    public static final Object read(final SChannelInput in) {
+    public final Object read(final VirtualFrame frame, final SChannelInput in) {
       try {
-        return in.read();
+        Object result = in.readAndSuspendWriter(afterWrite.executeCheckIsSetAndEnabled());
+        if (in.shouldBreakAfterRead()) {
+          haltNode.executeEvaluated(frame, result);
+        }
+        return result;
       } catch (InterruptedException e) {
         throw new RuntimeException(e);
+      }
+    }
+
+    @Override
+    protected boolean isTaggedWithIgnoringEagerness(final Class<?> tag) {
+      if (tag == ChannelRead.class || tag == ExpressionBreakpoint.class) {
+        return true;
+      } else {
+        return super.isTaggedWithIgnoringEagerness(tag);
       }
     }
   }
 
-  @Primitive(primitive = "procWrite:val:")
+  @Primitive(primitive = "procWrite:val:", selector = "write:")
   @GenerateNodeFactory
   public abstract static class WritePrim extends BinaryComplexOperation {
     @Child protected IsValue isVal;
 
+    /** Halt execution when triggered by breakpoint on write end. */
+    @Child protected UnaryExpressionNode haltNode;
+
+    /** Breakpoint info for triggering suspension after read. */
+    @Child protected AbstractBreakpointNode afterRead;
+
     public WritePrim(final boolean eagerlyWrapped, final SourceSection source) {
       super(eagerlyWrapped, source);
       isVal = IsValue.createSubNode();
+
+      haltNode = SuspendExecutionNodeGen.create(false, sourceSection, null);
+
+      if (VmSettings.TRUFFLE_DEBUGGER_ENABLED) {
+        Breakpoints bpCatalog = VM.getWebDebugger().getBreakpoints();
+        FullSourceCoordinate coord = SourceCoordinate.create(source);
+        this.afterRead = insert(BreakpointNodeGen.create(bpCatalog.getOppositeBreakpoint(coord)));
+      } else {
+        this.afterRead = insert(new DisabledBreakpointNode());
+      }
     }
 
     @Specialization
-    public final Object write(final SChannelOutput out, final Object val) {
+    public final Object write(final VirtualFrame frame, final SChannelOutput out,
+        final Object val) {
       if (!isVal.executeEvaluated(val)) {
         KernelObj.signalException("signalNotAValueWith:", val);
       }
       try {
-        out.write(val);
+        out.writeAndSuspendReader(val, afterRead.executeCheckIsSetAndEnabled());
+        if (out.shouldBreakAfterWrite()) {
+          haltNode.executeEvaluated(frame, val);
+        }
       } catch (InterruptedException e) {
         throw new RuntimeException(e);
       }
       return val;
+    }
+
+    @Override
+    protected boolean isTaggedWithIgnoringEagerness(final Class<?> tag) {
+      if (tag == ChannelWrite.class || tag == ExpressionBreakpoint.class) {
+        return true;
+      } else {
+        return super.isTaggedWithIgnoringEagerness(tag);
+      }
     }
   }